+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2026-09-14
+
+### Added
+
+- Leaderboard shows the difficulty of the best run.
+
+### Changed
+
+- Practice paddle is a bit wider, Overdrive a bit narrower.
+- HUD timer ticks once per second while the ball is in play.
+
 ## [0.5.1] - 2026-09-14
 
 ### Changed